Update main.js
Had a problem with a race condition - jQuery was loading after angular and 

    angular.element(document).ready(function() {
    		return require(["bootstrap"]);
    });

was never being fired - not sure why?? This caused the page to fail to render around 9 times out of 10.

Solved this by adding a dependency of jquery to lib/angular. 

I'm sure there is a better way to solve this but I needed to get it working quick! :)